# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Mady API
  description: |
    Mady（中观智能体）是一个面向专利/法律/IP 专业人士的智能 Agent 平台。

    ## 版本
    当前稳定版本 v1（2026-07-21）。所有端点通过 `/api/v1/*` 前缀标识。
    旧版 `/api/*` 路径保留向后兼容，但标记为已废弃（`deprecated: true`）。

    ## 认证
    技能管理端点（`/api/v1/skills/*`）支持可选的 Bearer Token 认证。
    配置 `SkillAPIAuthToken` 后，需要在请求头中携带 `Authorization: Bearer <token>`。

    ## 流式响应
    `/api/v1/chat` 支持两种模式：
    - `stream: false`（默认）— 返回 JSON
    - `stream: true` — 返回 SSE（Server-Sent Events）流

    ## 运维端点
    - `/health` — 存活检查（负载均衡）
    - `/ready` — 就绪检查（Agent 池 + Disclosure 健康度）
    - `/metrics` — Prometheus 格式指标
  version: "1.0.0"
  contact:
    name: Mady Authors
    url: https://github.com/xujian519/mady

servers:
  - url: http://localhost:8080
    description: 本地开发服务器

tags:
  - name: Chat
    description: 聊天与 Agent 交互
  - name: Threads
    description: 会话线程管理
  - name: Skills
    description: 技能管理
  - name: Disclosure
    description: 技术交底书分析
  - name: States
    description: 状态快照
  - name: AG-UI
    description: Agent GUI 事件流
  - name: Operations
    description: 运维与可观测性端点

paths:
  # ======================================================
  # 运维端点
  # ======================================================
  /health:
    get:
      tags: [Operations]
      summary: 存活检查
      description: 轻量级存活探测，适用于负载均衡器健康检查。返回固定 JSON 结构。
      responses:
        "200":
          description: 服务器存活
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /ready:
    get:
      tags: [Operations]
      summary: 就绪检查
      description: |
        检查服务器及其依赖组件是否就绪。
        - Agent 池：统计池中 Agent 数量
        - Disclosure 任务管理器：统计运行中的分析任务
      responses:
        "200":
          description: 服务器就绪
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"

  /metrics:
    get:
      tags: [Operations]
      summary: 监控指标
      description: |
        返回 Prometheus 格式的运行时指标：
        - `http_requests_total` — 请求总数
        - `http_request_errors_total` — 5xx 错误数
        - `http_active_connections` — 当前活跃连接数
      responses:
        "200":
          description: 指标文本
          content:
            text/plain:
              schema:
                type: string

  # ======================================================
  # Chat API (v1)
  # ======================================================
  /api/v1/chat:
    post:
      tags: [Chat]
      summary: 发送消息并获取 Agent 响应
      description: |
        主聊天端点。支持同步（JSON）和流式（SSE）两种模式。
        可携带 `thread_id` 实现多轮对话上下文保持。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: 同步响应（stream=false 时）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "200 (SSE)":
          description: 流式响应（stream=true 时）

  /api/chat:
    post:
      tags: [Chat]
      summary: 发送消息并获取 Agent 响应（已废弃）
      deprecated: true
      description: |
        旧版聊天端点。请迁移至 `/api/v1/chat`。
        主聊天端点。支持同步（JSON）和流式（SSE）两种模式。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: 同步响应
          content:
            application/json:
              schema:
                type: object
        "200 (SSE)":
          description: 流式响应

  # ======================================================
  # Skills API (v1)
  # ======================================================
  /api/v1/skills:
    get:
      tags: [Skills]
      summary: 列出已注册技能
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 技能列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"

  /api/skills:
    get:
      tags: [Skills]
      summary: 列出已注册技能（已废弃）
      deprecated: true
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 技能列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"

  /api/v1/skills/diagnostics:
    get:
      tags: [Skills]
      summary: 获取技能加载诊断信息
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 诊断报告

  /api/skills/diagnostics:
    get:
      tags: [Skills]
      summary: 获取技能加载诊断信息（已废弃）
      deprecated: true
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 诊断报告

  /api/v1/skills/events:
    get:
      tags: [Skills]
      summary: SSE 流式获取技能事件
      security:
        - BearerAuth: []
      responses:
        "200":
          description: SSE 事件流

  /api/skills/events:
    get:
      tags: [Skills]
      summary: SSE 流式获取技能事件（已废弃）
      deprecated: true
      security:
        - BearerAuth: []
      responses:
        "200":
          description: SSE 事件流

  /api/v1/skills/status:
    get:
      tags: [Skills]
      summary: 获取技能注册表状态
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 技能状态信息

  /api/skills/status:
    get:
      tags: [Skills]
      summary: 获取技能注册表状态（已废弃）
      deprecated: true
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 技能状态信息

  /api/v1/skills/reload:
    post:
      tags: [Skills]
      summary: 热重载技能（从磁盘重新扫描 SKILL.md）
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 重载完成
        "401":
          description: 未认证

  /api/skills/reload:
    post:
      tags: [Skills]
      summary: 热重载技能（已废弃）
      deprecated: true
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 重载完成
        "401":
          description: 未认证

  # ======================================================
  # Disclosure API (v1)
  # ======================================================
  /v1/disclosure/analyze:
    post:
      tags: [Disclosure]
      summary: 提交技术交底书，启动异步分析
      description: |
        接收交底书纯文本，后台运行 Pregel 分析管线（提取 → 一致性校验 →
        检索关键词 → 新颖性初判 → 报告生成 → 人工复核关卡）。
        返回任务 ID，通过轮询或 SSE 流获取进度与结果。
        任务结果保留 30 分钟后自动清理。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisclosureAnalyzeRequest"
      responses:
        "202":
          description: 任务已受理，返回任务 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisclosureAnalyzeResponse"
        "400":
          description: 请求体非法或 text 为空（返回 status=failed）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisclosureAnalyzeResponse"
        "500":
          description: 服务端未配置 LLM Provider（返回 status=failed）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisclosureAnalyzeResponse"

  /v1/disclosure/analyze/{task_id}:
    get:
      tags: [Disclosure]
      summary: 轮询分析任务状态与结果
      description: |
        查询指定任务的状态。完成时携带完整的分析报告和创造性评估结果。
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: 提交分析时返回的任务 ID
      responses:
        "200":
          description: 任务当前状态（完成时携带分析结果）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisclosureTaskStatus"
        "404":
          description: 任务不存在（或已过保留期被清理）
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /v1/disclosure/analyze/{task_id}/stream:
    get:
      tags: [Disclosure]
      summary: SSE 流式获取分析进度
      description: |
        以 Server-Sent Events 实时推送执行进度：
        - 连接建立后先发送 `: connected` 注释行
        - 每 500ms 推送一次 `event: progress`，data 为 DisclosureTaskStatus
          （含 progress，不含 result）
        - 任务结束时推送一次 `event: done`，data 为含 result/error 的
          最终 DisclosureTaskStatus，随后连接关闭
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: 提交分析时返回的任务 ID
      responses:
        "200":
          description: "SSE 事件流（event: progress / done）"
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/DisclosureTaskStatus"
        "404":
          description: 任务不存在（或已过保留期被清理）
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /v1/disclosure/analyze/{task_id}/review:
    post:
      tags: [Disclosure]
      summary: 提交人工复核结论
      description: |
        受理 awaiting_review 状态任务的人工复核结论。
        决策留痕到 ApprovalStore（P3 专家盲测数据收集触点）。
        仅在任务状态为 awaiting_review 时可用。
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: 待复核的任务 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisclosureReviewRequest"
      responses:
        "200":
          description: 复核结论已受理
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisclosureReviewResponse"
        "400":
          description: 请求体非法（decision 非 adopted/modified/rejected）
        "404":
          description: 任务不存在
        "409":
          description: 任务状态非 awaiting_review（无法复核）
        "503":
          description: ApprovalStore 未配置

  # ======================================================
  # Threads API (v1)
  # ======================================================
  /api/v1/threads:
    post:
      tags: [Threads]
      summary: 创建新线程
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: 线程已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
    get:
      tags: [Threads]
      summary: 列出所有线程
      responses:
        "200":
          description: 线程列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thread"

  /api/threads:
    post:
      tags: [Threads]
      summary: 创建新线程（已废弃）
      deprecated: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: 线程已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
    get:
      tags: [Threads]
      summary: 列出所有线程（已废弃）
      deprecated: true
      responses:
        "200":
          description: 线程列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thread"

  /api/v1/threads/{key}:
    get:
      tags: [Threads]
      summary: 获取线程详情
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 线程详情（含消息记录）
    delete:
      tags: [Threads]
      summary: 删除线程
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 已删除

  /api/threads/{key}:
    get:
      tags: [Threads]
      summary: 获取线程详情（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 线程详情
    delete:
      tags: [Threads]
      summary: 删除线程（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 已删除

  /api/v1/threads/{key}/branch:
    post:
      tags: [Threads]
      summary: 从当前叶子或指定节点创建分支
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: 新分支线程

  /api/threads/{key}/branch:
    post:
      tags: [Threads]
      summary: 从指定节点创建分支（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: 新分支线程

  /api/v1/threads/{key}/config:
    get:
      tags: [Threads]
      summary: 获取线程级调用配置
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 线程配置
    put:
      tags: [Threads]
      summary: 设置或清除线程级调用配置
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 配置已更新

  /api/threads/{key}/config:
    get:
      tags: [Threads]
      summary: 获取线程级调用配置（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 线程配置
    put:
      tags: [Threads]
      summary: 设置或清除线程级调用配置（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 配置已更新

  /api/v1/threads/{key}/thinking:
    get:
      tags: [Threads]
      summary: 获取线程级推理配置
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 推理配置
    put:
      tags: [Threads]
      summary: 设置或清除线程级推理配置
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 配置已更新

  /api/threads/{key}/thinking:
    get:
      tags: [Threads]
      summary: 获取线程级推理配置（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 推理配置
    put:
      tags: [Threads]
      summary: 设置或清除线程级推理配置（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 配置已更新

  # ======================================================
  # States API (v1)
  # ======================================================
  /api/v1/states:
    get:
      tags: [States]
      summary: 列出持久化状态键
      responses:
        "200":
          description: 状态键列表

  /api/states:
    get:
      tags: [States]
      summary: 列出持久化状态键（已废弃）
      deprecated: true
      responses:
        "200":
          description: 状态键列表

  /api/v1/states/{key}:
    get:
      tags: [States]
      summary: 获取状态快照
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 状态快照数据
    delete:
      tags: [States]
      summary: 删除状态快照
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 已删除

  /api/states/{key}:
    get:
      tags: [States]
      summary: 获取状态快照（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 状态快照数据
    delete:
      tags: [States]
      summary: 删除状态快照（已废弃）
      deprecated: true
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 已删除

  # ======================================================
  # AG-UI
  # ======================================================
  /agui/{path}:
    get:
      tags: [AG-UI]
      summary: Agent GUI 事件流处理器
      description: 将 Agent 执行过程通过 SSE 流式传输到 Web UI。
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE 事件流

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: 用于技能管理 API 的可选认证

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        time:
          type: string
          format: date-time
          example: "2026-07-21T12:00:00Z"

    ReadyResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        time:
          type: string
          format: date-time
        checks:
          type: object
          properties:
            agents:
              type: integer
              description: Agent 池总数量
            agents_idle:
              type: integer
              description: 空闲 Agent 数量
            disclosure_tasks:
              type: integer
              description: 运行中的 disclosure 分析任务数

    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: 用户消息内容
        thread_id:
          type: string
          description: 会话线程 ID（可选，用于保持上下文）
        stream:
          type: boolean
          default: false
          description: 是否使用 SSE 流式返回
        model:
          type: string
          description: 覆盖默认模型
        response_format:
          type: object
          description: 结构化输出格式配置
        thinking:
          type: object
          description: 推理过程配置
          properties:
            display:
              type: string
              enum: [collapsed, expanded, summarized, omitted]
            effort:
              type: string
              enum: [none, low, medium, high]
            budget:
              type: integer
              description: 推理 Token 预算
        skills:
          type: array
          items:
            type: string
          description: 启用的技能列表

    ChatResponse:
      type: object
      properties:
        output:
          type: string
          description: Agent 响应文本
        thread_id:
          type: string
          description: 会话线程 ID
        error:
          type: string
          description: 错误信息（失败时存在）

    Skill:
      type: object
      properties:
        name:
          type: string
          description: 技能名称
        description:
          type: string
          description: 技能描述
        domain:
          type: string
          description: 所属领域

    Thread:
      type: object
      properties:
        id:
          type: string
          description: 线程唯一标识
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer

    DisclosureAnalyzeRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: 技术交底书纯文本内容（非空）

    DisclosureAnalyzeResponse:
      type: object
      properties:
        task_id:
          type: string
          description: 任务 ID（格式 disclosure-{时间戳}-{序号}）；失败时为空
        status:
          type: string
          description: 受理成功为 pending；请求被拒绝为 failed

    DisclosureReviewRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [adopted, modified, rejected]
          description: 复核决定：adopted（采纳）/ modified（修改后采纳）/ rejected（拒绝）
        modified_output:
          type: string
          description: modified 决策时的改后内容
        feedback:
          type: string
          description: 复核反馈意见
        case_id:
          type: string
          description: 关联案件 ID

    DisclosureReviewResponse:
      type: object
      properties:
        task_id:
          type: string
          description: 任务 ID
        status:
          type: string
          description: 受理后状态（reviewed）

    DisclosureTaskStatus:
      type: object
      properties:
        task_id:
          type: string
          description: 任务 ID
        status:
          type: string
          enum: [pending, running, awaiting_review, reviewed, completed, failed]
          description: |
            任务状态：pending 排队中 / running 执行中 /
            awaiting_review 分析完成待人工复核 /
            reviewed 已人工复核 / completed 已完成 / failed 失败
        progress:
          $ref: "#/components/schemas/DisclosureProgress"
        result:
          $ref: "#/components/schemas/DisclosureReport"
        error:
          type: string
          description: 失败原因（status=failed 时存在）
        review_decision:
          type: string
          description: 人工复核决定（adopted/modified/rejected）
        inventiveness:
          type: object
          description: 创造性三步法评估结果

    DisclosureProgress:
      type: object
      properties:
        current_node:
          type: string
          description: 当前执行中的 Pregel 节点
        nodes_done:
          type: array
          items:
            type: string
          description: 已完成节点列表

    DisclosureReport:
      type: object
      description: 结构化分析报告（disclosure.AnalysisReport）
      properties:
        id:
          type: string
          description: 报告 ID
        document:
          type: object
          description: 解析后的交底书文档（标题/格式/段落）
        extraction:
          type: object
          description: 提取结果（技术问题/技术特征/技术效果/PFE 三元组）
        consistency:
          type: object
          description: 一致性校验结果（得分/问题列表/重试状态）
        search_keywords:
          type: array
          items:
            type: string
          description: 推荐检索关键词
        novelty:
          type: object
          description: 新颖性初判（assessed/conclusion/notes）
        report_text:
          type: string
          description: LLM 生成的报告正文（Markdown）
        generated_at:
          type: string
          format: date-time
        reviewed_by_human:
          type: boolean
          description: 是否已通过人工复核